Added File2Deck parsing, implemented deck draw
Reworked some systems, added file parsing for deck loading, and added
random card drawing.  This project has a lot of ugly code (like string
to enum conversions and object creation functions and the Card's copy
function) but if I really need to fix it, it's still reasonably sane
refactoring.

None of it is particularly infused in the design, so it should be quick
to pull parts and swap as needed.