Trying to prep for adding client/server
Tl;dr Pushing commit so I know where I need to work.

A bunch of stuff going on, it's kind of a mess right now.  This isn't
really a great commit in the sense that stuff is half done, but I'd
rather not let it all pile up.  All Packet types have been tested for
AsByteArray and FromByteArray.  These also validate that ByteArrayReader
is working.